+++ src/esdt/change-properties-esdt.ts
@@ -3,7 +3,6 @@ import { exit } from 'process';
 import {
   Transaction,
   BytesValue,
-  AddressValue,
   Address,
   ContractCallPayloadBuilder,
   ContractFunction,
@@ -16,12 +15,10 @@ import {
   shortChainId,
   builtInSC,
   commonOpertationsGasLimit,
-  esdtTokenSpecialRoles,
+  esdtTokenProperties,
 } from '../config';
 
-type OperationType = 'set' | 'unset';
-
-const promptQuestions = (type: OperationType): PromptObject[] => [
+const promptQuestions: PromptObject[] = [
   {
     type: 'text',
     name: 'ticker',
@@ -29,34 +26,22 @@ const promptQuestions = (type: OperationType): PromptObject[] => [
     validate: (value) => (!value ? 'Required!' : true),
   },
   {
-    type: 'text',
-    name: 'address',
-    message: `Please provide the address ${
-      type === 'set' ? 'to assign' : 'with'
-    } the role.\n`,
-    validate: (value) => (!value ? 'Required!' : true),
-  },
-  {
     type: 'multiselect',
-    name: 'specialRoles',
-    message: `Please choose special roles to ${
-      type === 'set' ? 'assign' : 'remove'
-    }.\n`,
-    choices: esdtTokenSpecialRoles.map((property) => ({
+    name: 'tokenProperties',
+    message: `Please choose a new set of the properties for the token.\n`,
+    choices: esdtTokenProperties.map((property) => ({
       title: property,
       value: property,
     })),
   },
 ];
 
-export const toggleSpecialRolesEsdt = async (type: OperationType) => {
+export const changePropertiesEsdt = async () => {
   try {
-    const { ticker, address, specialRoles } = await prompts(
-      promptQuestions(type)
-    );
+    const { ticker, tokenProperties } = await prompts(promptQuestions);
 
-    if (!ticker || !address) {
-      console.log('You have to provide the ticker and address!');
+    if (!ticker) {
+      console.log('You have to provide the ticker!');
       exit(9);
     }
 
@@ -64,23 +49,21 @@ export const toggleSpecialRolesEsdt = async (type: OperationType) => {
 
     const { signer, userAccount, provider } = await setup();
 
-    const args: TypedValue[] = [
-      BytesValue.fromUTF8(ticker),
-      new AddressValue(new Address(address.trim())),
-    ];
+    const args: TypedValue[] = [BytesValue.fromUTF8(ticker)];
 
-    for (const role of esdtTokenSpecialRoles) {
-      if (specialRoles.includes(role)) {
-        args.push(BytesValue.fromUTF8(role));
+    for (const property of esdtTokenProperties) {
+      let propertyEnabled = false;
+
+      if (tokenProperties.includes(property)) {
+        propertyEnabled = true;
       }
+
+      args.push(BytesValue.fromUTF8(property));
+      args.push(BytesValue.fromUTF8(propertyEnabled.toString()));
     }
 
     const data = new ContractCallPayloadBuilder()
-      .setFunction(
-        new ContractFunction(
-          type === 'set' ? 'setSpecialRole' : 'unSetSpecialRole'
-        )
-      )
+      .setFunction(new ContractFunction('controlChanges'))
       .setArgs(args)
       .build();